rename fix assesments upsell model content
