+++ packages/js/src/shared-admin/components/index.js
@@ -6,4 +6,4 @@ export { OutboundLink } from "./outbound-link";
 export { PremiumUpsellCard } from "./premium-upsell-card";
 export { RecommendationsSidebar } from "./recommendations-sidebar";
 export { VideoFlow } from "./video-flow";
-export { AiFixAssesmentsUpsell } from "./ai-fix-assesments";
+export { AiFixAssesmentsUpsell } from "./ai-fix-assesments-upsell";
